requirements.txt
aiogram==3.3.0
httpx==0.27.0
python-dotenv==1.0.0
numpy==1.26.4
Pillow==10.4.0
deep-translator==1.11.4
requests==2.32.3
opencv-python-headless==4.10.0.84
fastapi==0.110.0
uvicorn==0.27.1